fix timer initialisation for background task
If using with the Arduino environment, the timers are already set to PWM
mode. This means OCRx is a double-buffered register. The written value isn't
used before the next OC event. If the mode is changed to CTC, this update
from the shadow register never happens and the written value never come into
effect.

To avoid this, change the mode before writing a new value to OCRx.