Use proper fixed point multiplication for the control loop
use the special s16=s16*u8.8 multiplication function from the original
assembler version. This multiplication includes a right shift by 8 bits to
convert the s24:8 result to s16.

It would be possible to use the build-in long arithmetics, but that compiles
terribly ineffienct and is even slower than using this software
multiplication.

Using the AVR hardware multiplication unit would be possible, but is not
implemented yet.